+++ ios/Plugin/Utils/UtilsMigrate.swift
@@ -9,6 +9,7 @@ enum UtilsMigrateError: Error {
     case addSQLiteSuffix(message: String)
     case getMigratableList(message: String)
     case deleteOldDatabases(message: String)
+    case moveDatabasesAndAddSuffix(message: String)
 }
 
 class UtilsMigrate {
@@ -187,6 +188,77 @@ class UtilsMigrate {
             var msg: String = "addSQLiteSuffix command failed :"
             msg.append(" \(error.localizedDescription)")
             throw UtilsMigrateError.addSQLiteSuffix(message: msg)
+        }
+    }
+    // swiftlint:enable cyclomatic_complexity
+    // swiftlint:enable function_body_length
+
+    // MARK: - moveDatabasesAndAddSuffix
+
+    // swiftlint:disable function_body_length
+    // swiftlint:disable cyclomatic_complexity
+    class func moveDatabasesAndAddSuffix(databaseLocation: String, folderPath: String,
+                                         dbList: [String]) throws {
+        var fromFile: String = ""
+        var toFile: String = ""
+        do {
+            let databaseURL: URL = try UtilsFile
+                .getFolderURL(folderPath: databaseLocation)
+            let dbPathURL: URL = try UtilsFile
+                .getFolderURL(folderPath: folderPath)
+            var isDir = ObjCBool(true)
+            if FileManager.default.fileExists(atPath: dbPathURL.relativePath,
+                                              isDirectory: &isDir) &&
+                isDir.boolValue {
+                var mDbList: [String]
+                if dbList.count > 0 {
+                    mDbList = try UtilsFile
+                        .getFileList(path: dbPathURL.relativePath, ext: nil)
+                } else {
+                    mDbList = try UtilsFile
+                        .getFileList(path: dbPathURL.relativePath, ext: "db")
+                }
+                for file: String in mDbList {
+                    if !file.contains("SQLite.db") {
+                        fromFile = file
+                        toFile = ""
+                        if dbList.count > 0 {
+                            if dbList.contains(fromFile) {
+                                if String(file.suffix(3)) == ".db" {
+                                    toFile = file
+                                        .replacingOccurrences(of: ".db", with: "SQLite.db")
+                                } else {
+                                    toFile = file + "SQLite.db"
+                                }
+                            }
+                        } else {
+                            toFile = file
+                                .replacingOccurrences(of: ".db", with: "SQLite.db")
+                        }
+                        if !toFile.isEmpty {
+                            let uFrom: URL = dbPathURL.appendingPathComponent(fromFile)
+                            let uTo: URL = databaseURL.appendingPathComponent(toFile)
+                            try UtilsFile.moveFile(pathName: uFrom.path, toPathName: uTo.path, overwrite: true)
+                        }
+                    }
+                }
+                return
+            } else {
+                var msg: String = "moveDatabasesAndAddSuffix command failed :"
+                msg.append(" Folder '\(dbPathURL.absoluteString)' not found")
+                throw UtilsMigrateError.moveDatabasesAndAddSuffix(message: msg)
+            }
+        } catch UtilsFileError.getDatabasesURLFailed {
+            throw UtilsMigrateError
+            .moveDatabasesAndAddSuffix(message: "getDatabasesURLFailed")
+        } catch UtilsFileError.getFolderURLFailed(let message) {
+            throw UtilsMigrateError.moveDatabasesAndAddSuffix(message: message)
+        } catch UtilsFileError.getFileListFailed {
+            throw UtilsMigrateError.moveDatabasesAndAddSuffix(message: "getFileListFailed")
+        } catch let error {
+            var msg: String = "moveDatabasesAndAddSuffix command failed :"
+            msg.append(" \(error.localizedDescription)")
+            throw UtilsMigrateError.moveDatabasesAndAddSuffix(message: msg)
         }
     }
     // swiftlint:enable cyclomatic_complexity